ocr-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
requests==2.31.0
pillow==10.2.0
# torch
# transformers
